backend/requirements.txt
# Backend dependencies

# ==================== FRAMEWORK ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1

# ==================== DATABASE ====================
supabase==2.4.0
psycopg2-binary==2.9.9

# ==================== CACHE ====================
redis==5.0.1

# ==================== AUTH ====================
PyJWT==2.12.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.5

# ==================== HTTP ====================
requests==2.31.0

# ==================== AI ====================
google-generativeai==0.3.0

# ==================== STORAGE ====================
boto3==1.34.21

# ==================== TESTING ====================
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# ==================== DEVELOPMENT ====================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
python-dotenv==1.0.0
